requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
SQLAlchemy==2.0.40
alembic==1.15.2
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
Jinja2==3.1.6
httpx==0.28.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-xdist==3.6.1
flake8==7.2.0
black==25.1.0
isort==6.0.1